refactor: convert isImageUrl to utils
